Adds README, .python-version, and requirements.txt.
These two files describe our development environment. This makes it
really easy to set such an environment up for anybody who wishes to
contribute to the project.

I've added a README.md that we can use to describe what the project is
and provides steps to setting up an environment.